Implement core functionality with error handling, logging, etc.

+++ src/core/config.py
@@ -22,7 +22,7 @@ class GeoConfig:
         "http://ip-api.com/json",
     )
 
-    TIMEOUT: float = float(
+    TIMEOUT: int = int(
         os.getenv(
             "IP_API_TIMEOUT",
             10,
